Sorts grammar by their size
This prevents, in most cases, having to use things like lookahead
characters. For example, the grammar definitions of value and expression
overlap, so the compiler will look to see if a set of tokens better
represents an expression, since the expression is the longer type. If
this fails, it matches tokens against the value grammar.